move expired objects delete out of gtx
