+++ hugegraph-test/src/main/resources/hugegraph.properties
@@ -1,7 +1,7 @@
 gremlin.graph=com.baidu.hugegraph.HugeFactory
 
-backend=cassandra
-serializer=cassandra
+backend=${backend}
+serializer=${serializer}
 
 store=hugegraph
